[Refactor/issue-254] PerformanceCard를 컴파운드로 제공 (#264)
* rename: PerformanceCard  => PerformancCardHeadless

* feat: compound 패턴의 PerformanceCard 추가 (=== MianPerformanceCard)

* chore: 의존 경로 변경
