+++ src/components/common/PerformanceCardHeadless/index.ts
@@ -1,0 +1,25 @@
+import {
+  Root,
+  Image,
+  Title,
+  Status,
+  DateRange,
+  Location,
+  Cast,
+  Price,
+  LikeButton,
+} from './PerformanceCardHeadless';
+
+const PerformanceCardHeadless = {
+  Root,
+  Image,
+  Title,
+  Status,
+  DateRange,
+  Location,
+  Cast,
+  Price,
+  LikeButton,
+};
+
+export default PerformanceCardHeadless;
