+++ src/components/pages/performances/index.ts
@@ -1,6 +1,5 @@
 export { default as PerformanceList } from './PerformanceList';
 export { default as PerformanceListContainer } from './PerformancesListContainer';
-export { default as PerformanceCard } from './PerformancesPerformanceCard';
 export { default as Search } from './PerformancesSearchInput';
 export { default as Filter } from './PerformancesFilter';
 export { default as SortDropdown } from './PerformancesSortDropdown';
